+++ src/lib/reporters/phpunit/bootstrap.php
@@ -2,11 +2,16 @@
 
 try {
     $pattern = '/^lode_bootstrap=/i';
-    $bootstrap = array_values(preg_grep($pattern, $_SERVER['argv']))[0];
+    $bootstrap = isset(array_values(preg_grep($pattern, $_SERVER['argv']))[0]) ? array_values(preg_grep($pattern, $_SERVER['argv']))[0] : false;
     if (!$bootstrap) {
-        throw new Exception;
+        // If we've already bootstrapped Lode before, we should be able to
+        // retrieve the bootstrap location again. Otherwise, fail.
+        $bootstrap = file_get_contents(__DIR__ . DIRECTORY_SEPARATOR . 'bootstrap', $bootstrap);
+        if ($bootstrap === false) {
+            throw new Exception;
+        }
     }
-} catch (\Exception $e) {
+} catch (Exception $e) {
     echo('Unable to locate bootstrap file. Have you configured it in your Lode framework settings?');
     exit(1);
 }
@@ -27,8 +32,13 @@ spl_autoload_register(function ($class) {
             }
         }
 
-        include_once __DIR__ . DIRECTORY_SEPARATOR. 'src' . $folder . str_replace('\\', DIRECTORY_SEPARATOR, substr($class, strlen('LodeApp\PHPUnit')) . '.php');
+        include_once __DIR__ . DIRECTORY_SEPARATOR . 'src' . $folder . str_replace('\\', DIRECTORY_SEPARATOR, substr($class, strlen('LodeApp\PHPUnit')) . '.php');
     }
 });
 
+// Remember actual bootstrap location in case PHPUnit boots another process
+// for tests running in isolation.
+file_put_contents(__DIR__ . DIRECTORY_SEPARATOR . 'bootstrap', $bootstrap);
+
+// Now that Lode is setup, require the original user-defined bootstrap.
 require_once preg_replace($pattern, '', $bootstrap);