+++ src/lib/reporters/phpunit/src/48-57/Report.php
@@ -11,21 +11,6 @@ use Throwable;
 class Report
 {
     /**
-     * A map of PHPUnit vs. Lode string statuses
-     *
-     * @var array
-     */
-    protected $statuses = [
-        0 => 'passed', // PASSED
-        1 => 'skipped', // SKIPPED
-        2 => 'incomplete', // INCOMPLETE
-        3 => 'failed', // FAILURE
-        4 => 'failed', // ERROR
-        5 => 'empty', // RISKY
-        6 => 'warning', // WARNING
-    ];
-
-    /**
      * Name prettifier, for consistent transformations.
      *
      * @var \PHPUnit_Util_TestDox_NamePrettifier
@@ -46,6 +31,14 @@ class Report
      * @var PHPUnit_Framework_Test
      */
     protected $test;
+
+    /**
+     * The outcome of the test run, standardised
+     * for display inside of Lode.
+     *
+     * @var string
+     */
+    protected $status;
 
     /**
      * The exception triggered by the test, if any.
@@ -86,10 +79,12 @@ class Report
      * Create a new Lode report class.
      *
      * @param \PHPUnit_Framework_Test $test
+     * @param string $status
      */
-    public function __construct(PHPUnit_Framework_Test $test)
+    public function __construct(PHPUnit_Framework_Test $test, $status = 'idle')
     {
         $this->test = $test;
+        $this->status = $status;
         $this->class = get_class($this->test);
         if ($this->isWarning()) {
             $original = $this->getWarningClass();
@@ -260,7 +255,7 @@ class Report
     public function render()
     {
         $current = array_merge($this->hydrateTest(), [
-            'status' => $this->status(),
+            'status' => $this->status,
             'feedback' => $this->exception ? (new Feedback($this->exception))->render() : null,
             'console' => $this->console->pullTestLogs($this->getFileName(), $this->test->getName()),
             'stats' => [
@@ -298,17 +293,6 @@ class Report
     protected function transformName(string $name)
     {
         return preg_replace('/^it\s/', '', $this->prettifier->prettifyTestMethod($name));
-    }
-
-    /**
-     * Map a PHPUnit status string into a standardized
-     * Lode status string.
-     *
-     * @return string
-     */
-    protected function status()
-    {
-        return Util::get($this->statuses, $this->test->getStatus(), 'warning');
     }
 
     /**