+++ support/release-notes.md
@@ -1,2 +1,5 @@
+Changed:
+    - Improved parsing of Jest failure feedback
+    - Always use panels for diff
 Fixed:
-    - Security fixes - #10 #11
+    - PHPUnit unable to run tests in separate processes